# ARROW-10366: [Rust][DataFusion] Do not buffer intermediate results in merge or HashAggregate