@andygrove 2020-10-30
I will try and run some TPC benchmarks tomorrow. I'm excited to see how this looks. Great work @jorgecarleitao and @alamb