@alamb 2020-10-31
> I went through this. Beautiful work, @alamb !

Thanks @jorgecarleitao 

> My only comment is that we are performing a change in the git submodule testing which I do not think it was intended.

You are correct that that was not intended -- I will fix that