@andygrove 2020-10-31
Here are the TPC benchmark results (sf=100) based on the following command.

```bash
cargo run --release --bin tpch -- --iterations 3 --path /mnt/tpch/parquet-100GB --format parquet --query 1 --batch-size 4096 --concurrency 24
```

### 2.0.0 release

```
Query 1 iteration 0 took 13779 ms
Query 1 iteration 1 took 13665 ms
Query 1 iteration 2 took 14008 ms
```

### master branch

```
Query 1 iteration 0 took 20304 ms
Query 1 iteration 1 took 20228 ms
Query 1 iteration 2 took 20737 ms
```

### This PR

```
Query 1 iteration 0 took 20528 ms
Query 1 iteration 1 took 20688 ms
Query 1 iteration 2 took 20432 ms
```

I'm not sure why there is such a regression since 2.0.0 and will see if I can track down where this started.