@andygrove 2020-10-31
The performance regression is related to removing use of the specialization trait. I filed https://issues.apache.org/jira/browse/ARROW-10453