@alamb 2020-11-01
@andygrove 

Do you have the instructions for generating (or would be willing to provide me a copy to) the 100GB parquet dataset you are using in the following benchmark? I would love to try and run it locally and do some profiling to understand more of where time is being spent.

```
cargo run --release --bin tpch -- --iterations 3 --path /mnt/tpch/parquet-100GB --format parquet --query 1 --batch-size 4096 --concurrency 24
```

Thanks!
Andrew