@andygrove 2020-11-01
@alamb I've been using Apache Spark to do the conversion and re-partitioning, using code in the Ballista project that I've been running manually from my IDE. I wouldn't feel good about checking Scala code into the Arrow Rust project so maybe the short term solution is for me to put a Dockerfile in Ballista for creating a Docker image that can do the conversion. Long term we should write a Rust utility to read CSV and write to Parquet. I'm not sure if the Parquet writer is far enough along yet to support that?