@andygrove 2020-11-01
@alamb Could you try this out. You'll need to map volumes in appropriately to read/write from host directories.

```bash
docker run -it ballistacompute/spark-benchmarks:0.4.0-SNAPSHOT 
  -h, --help   Show help message

Subcommand: convert-tpch
  -i, --input  <arg>
      --input-format  <arg>
  -o, --output  <arg>
      --output-format  <arg>
  -p, --partitions  <arg>
  -h, --help                   Show help message
Subcommand: tpch
      --input-format  <arg>
  -i, --input-path  <arg>
  -q, --query  <arg>
  -h, --help                  Show help message
```