@alamb 2020-11-02
> @alamb Could you try this out. You'll need to map volumes in appropriately to read/write from host directories.

Checking it out...