@alamb 2020-11-02
@andygrove  -- it seems to work well:

```
$ docker run  -it -v /Users/alamb/Software/tpch_data/SF1:/opt/spark/work-dir ballistacompute/spark-benchmarks:0.4.0-SNAPSHOT convert-tpch  --input lineitem.tbl --input-format tbl --output lineitem.parquet --output-format parquet --partitions 16
```

Which then produced output like this:

```
$ ls -l /Users/alamb/Software/tpch_data/SF1/lineitem.parquet/
total 428032
-rw-r--r--  1 alamb  staff     0B Nov  2 06:27 _SUCCESS
-rw-r--r--  1 alamb  staff    13M Nov  2 06:27 part-00000-47e26762-3d8a-4652-ba5e-2e2a620e399e-c000.snappy.parquet
-rw-r--r--  1 alamb  staff    13M Nov  2 06:27 part-00001-47e26762-3d8a-4652-ba5e-2e2a620e399e-c000.snappy.parquet
-rw-r--r--  1 alamb  staff    13M Nov  2 06:27 part-00002-47e26762-3d8a-4652-ba5e-2e2a620e399e-c000.snappy.parquet
-rw-r--r--  1 alamb  staff    13M Nov  2 06:27 part-00003-47e26762-3d8a-4652-ba5e-2e2a620e399e-c000.snappy.parquet
-rw-r--r--  1 alamb  staff    13M Nov  2 06:27 part-00004-47e26762-3d8a-4652-ba5e-2e2a620e399e-c000.snappy.parquet
-rw-r--r--  1 alamb  staff    13M Nov  2 06:27 part-00005-47e26762-3d8a-4652-ba5e-2e2a620e399e-c000.snappy.parquet
-rw-r--r--  1 alamb  staff    13M Nov  2 06:27 part-00006-47e26762-3d8a-4652-ba5e-2e2a620e399e-c000.snappy.parquet
-rw-r--r--  1 alamb  staff    13M Nov  2 06:27 part-00007-47e26762-3d8a-4652-ba5e-2e2a620e399e-c000.snappy.parquet
-rw-r--r--  1 alamb  staff    13M Nov  2 06:27 part-00008-47e26762-3d8a-4652-ba5e-2e2a620e399e-c000.snappy.parquet
-rw-r--r--  1 alamb  staff    13M Nov  2 06:27 part-00009-47e26762-3d8a-4652-ba5e-2e2a620e399e-c000.snappy.parquet
-rw-r--r--  1 alamb  staff    13M Nov  2 06:27 part-00010-47e26762-3d8a-4652-ba5e-2e2a620e399e-c000.snappy.parquet
-rw-r--r--  1 alamb  staff    13M Nov  2 06:27 part-00011-47e26762-3d8a-4652-ba5e-2e2a620e399e-c000.snappy.parquet
-rw-r--r--  1 alamb  staff    13M Nov  2 06:27 part-00012-47e26762-3d8a-4652-ba5e-2e2a620e399e-c000.snappy.parquet
-rw-r--r--  1 alamb  staff    13M Nov  2 06:27 part-00013-47e26762-3d8a-4652-ba5e-2e2a620e399e-c000.snappy.parquet
-rw-r--r--  1 alamb  staff    13M Nov  2 06:27 part-00014-47e26762-3d8a-4652-ba5e-2e2a620e399e-c000.snappy.parquet
-rw-r--r--  1 alamb  staff    13M Nov  2 06:27 part-00015-47e26762-3d8a-4652-ba5e-2e2a620e399e-c000.snappy.parquet
```

I plan to test out using these files with datafusion later today if I have time, or tomorrow if not. 

Thanks again,
Andrew