@andygrove 2020-11-05
Thanks @alamb for posting the detailed info. I hope I haven't caused confusion with my earlier comments and performance data. I was not seeing a slowdown on this PR compared to master. I was seeing a slowdown in master compared to the 2.0.0 release (which I tracked down to the commit that removed specialization).